Only autoraster if plot fmt is specified
Default line can handle huge number of datapoints while scatter plots
can't and need rasterisation to be practicable. Could be improved by
checking for a scatter-type format character in the fmt kwarg or third
arg, but this approach is simple enough and not important enough to
warrant the time expenditure.